add select style, move tweet under search (dir only)
