+++ package.json
@@ -15,9 +15,6 @@
   "devDependencies": {
     "@babel/core": "^7.7.4",
     "@babel/plugin-proposal-class-properties": "^7.7.4",
-    "@babel/plugin-proposal-export-default-from": "^7.10.4",
-    "@babel/plugin-proposal-optional-chaining": "^7.11.0",
-    "@babel/plugin-syntax-export-default-from": "^7.10.4",
     "@babel/preset-env": "^7.7.4",
     "@babel/preset-react": "^7.7.4",
     "babel-eslint": "^10.0.3",
@@ -54,6 +51,8 @@
   },
   "dependencies": {
     "@appbaseio/reactivesearch": "^3.1.1",
+    "@babel/plugin-proposal-export-default-from": "^7.10.4",
+    "@babel/plugin-proposal-optional-chaining": "^7.11.0",
     "apexcharts": "^3.10.1",
     "bodybuilder": "^2.2.18",
     "date-fns": "^2.9.0",
