+++ src/pages/search/search.style.scss
@@ -29,4 +29,8 @@
       padding: 10px;
     }
   }
+
+  select[name="sort-options"] {
+    border: 3px solid black;
+  }
 }
